requirements.txt
selenium
quart
Pillow
